Remove the common request method to module util
